pyproject.toml
[project]
name = "cali"
version = "0.1.0"
description = "Screenshot your schedule and this program will add it to google calendar."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "google-auth-oauthlib>=1.2.2",
    "openai>=1.102.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "google-genai",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.181.0",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "google-cloud-vision>=3.10.2",
]
